relocate component ErrorCatcher, edit height oh image
