+++ src/app/(site)/error.tsx
@@ -1,0 +1,13 @@
+'use client';
+import { useEffect } from 'react';
+
+import { ErrorCatcher } from '@/sections/ErrorCatcher';
+
+const Error = ({ error }: { error: Error & { digest?: string } }) => {
+  useEffect(() => {
+    console.log(error);
+  }, [error]);
+
+  return <ErrorCatcher error="server-error" />;
+};
+export default Error;
