+++ src/sections/ErrorCatcher/types.ts
